net: vlan: Use vlan_prio instead of vlan_qos in mapping
The vlan_qos member is used to save the vlan qos, but we only save the
priority. Also, we will get the priority in vlan netlink and proc.
We can just save the vlan priority using vlan_prio, so we can use vlan_prio
to get the priority directly.

For flexibility, we introduced vlan_dev_get_egress_priority() helper
function. After this patch, we will call vlan_dev_get_egress_priority()
instead of vlan_dev_get_egress_qos_mask() in irdma.ko and rdma_cm.ko.
Because we don't need the shift and mask operations anymore.

There is no functional changes.

+++ net/8021q/vlan.h
@@ -126,9 +126,9 @@ void vlan_filter_drop_vids(struct vlan_info *vlan_info, __be16 proto);
 
 /* found in vlan_dev.c */
 void vlan_dev_set_ingress_priority(const struct net_device *dev,
-				   u32 skb_prio, u16 vlan_prio);
+				   u32 skb_prio, u8 vlan_prio);
 int vlan_dev_set_egress_priority(const struct net_device *dev,
-				 u32 skb_prio, u16 vlan_prio);
+				 u32 skb_prio, u8 vlan_prio);
 void vlan_dev_free_egress_priority(const struct net_device *dev);
 int vlan_dev_change_flags(const struct net_device *dev, u32 flag, u32 mask);
 void vlan_dev_get_realdev_name(const struct net_device *dev, char *result,
